use netlink library to interface with nftables
kube-network-policies only need a specific set of nftables rules
to be present to filter the undesired traffic and enqueue the traffic
subject to inspection to userspace.

There is an optimization using sets to avoid diverting the traffic that
are not subject to policies to avoid the penalty of userspace, but
besides that there is no plans to require more interaction with
netfilter.

Using the netfilter interface with nftables is complex and very low
level, but since the netfilter interaction should not change much in the
foreseeble future, the removal of the dependency on the userspace tools
bring a big advantage in term of image size 72MB vs 92MB as today, and
n the maintanance of the image, since we only need to maintain the
golang binary.