switch to nfqueue
nflog is great to capture the packets because it does not participate
on the datapath so there is no risk of causing problems. However, it
turns out nflog is too slow, and when the dns packets are processed by
the agent, the client already sent the TCP request, so it is not
possible to get the resolved IPs on time.

Using nfqueue allow us to hold the packet, guaranteeing that we are
going to have the resolved addresses  before the client sends the
request.