obtain PodIPs locally
there is a known race if we depend on the published podIPs, because
the pod is created by the runtime, the info is passed to the kubelet and
then kubelet updates the pod object. This change sends an update to the
agent via the watch, and this roundtrip is racy.
NRI runs local on the node, and can be used to obtain the assigned pod
ips before the pod is running, supressing this race.