add ipcache
use ipcache for a decoupled architecture.
There is a new component that watches the necessary information and node
agents just connect to that component/s.

It uses etcd to store and server the data.
The client uses bboltdb to avoid having all data in memory and uses a
LRU cache for fast retrieval. It uses etcd watchers to keep the local db
in sync.

With this model the clients only need to watch the networkPolicy
objects.