move to a pluggable architecture
golang modules are not easy to work with as the modules need to match
the dependencies, creating a tightly coupling.

Since the feature set is not big and we just want to get a minimal
decoupling to be able to develop the new features in the Network Policy
Working Group withoutbreaking the existing consumers and user of the
project, we just create multiple binaries with the features required.
Unstable features will live under the plugins/ folder and moved once
they are stable and GA to the main binary.