recalculate network policies for established connections
is it possible for a network policy or namespace or pod labels to change
during the lifetime of this objects.

Since these changes can impact the policy over the connectons already
established we need to reevaluate them, in order to do that, we just
check the existing conntrack entries and we clean them to force the
subsequent packets to go over the nfqueue process for non TCP
connections.

For TCP connections we try to find the socket and kill it.