make firewall enforcer for current connections optional
besides we discussed and deccided that is the more logical behavior, the
previous behavior of not impacting existing connections is widely used
even outside of kubernetes, so in order to avoid breaking existing
behavior we rollout this new functionality behind a flag that allow
users to define the period of time they want to use to enforce the
network policies on existing connections. The default value is zero,
that disables this feature.