enable network  policies strict mode
apply network policies to existing connections.

The dataplane now inspect the existing connections in the conntrack
table and evaluates against the current network policies.
If a established connection is no longer allowed then the dataplane:
For UDP it flush the conntrack entry, next packet will be reevaluated.
For TCP it creates an nftables Set with the 4-tuple, and the packets
matching that 4 tuple are enqueued, so we do no have the risk of
dropping packets that may be valid later.